+++ src/Page.php
@@ -388,11 +388,12 @@ class Page
      *      ->saveToFile('/tmp/image.jpg');
      * ```
      *
+     * @param int|null $timeout
      * @return Clip
      */
-    public function getFullPageClip(): Clip
+    public function getFullPageClip(int $timeout = null): Clip
     {
-        $contentSize = $this->getLayoutMetrics()->await()->getContentSize();
+        $contentSize = $this->getLayoutMetrics()->await($timeout)->getContentSize();
         return new Clip(0, 0, $contentSize['width'], $contentSize['height']);
     }
 
@@ -497,6 +498,16 @@ class Page
      * @param array $options
      *  - landscape: default false
      *  - printBackground: default false
+     *  - displayHeaderFooter: default false
+     *  - headerTemplate: HTML template for the print header (see docs for details)
+     *  - footerTemplate: HTML template for the print footer (see docs for details)
+     *  - paperWidth: default 8.5 inches
+     *  - paperHeight: default 11 inches
+     *  - marginTop: default 1 cm
+     *  - marginBottom: default 1 cm
+     *  - marginLeft: default 1 cm
+     *  - marginRight: default 1 cm
+     *  - preferCSSPageSize: default false
      * @return PagePdf
      * @throws CommunicationException
      */
@@ -527,16 +538,38 @@ class Page
             }
             $pdfOptions['printBackground'] = $options['printBackground'];
         }
-
+    
         // option displayHeaderFooter
         if (array_key_exists('displayHeaderFooter', $options)) {
-            // printBackground requires type to be boolean
+            // displayHeaderFooter requires type to be boolean
             if (!is_bool($options['displayHeaderFooter'])) {
                 throw new \InvalidArgumentException(
                     'Invalid options "displayHeaderFooter" for print to pdf. Must be true or false'
                 );
             }
             $pdfOptions['displayHeaderFooter'] = $options['displayHeaderFooter'];
+        }
+    
+        // option headerTemplate
+        if (array_key_exists('headerTemplate', $options)) {
+            // headerTemplate requires type to be string
+            if (!is_string($options['headerTemplate'])) {
+                throw new \InvalidArgumentException(
+                    'Invalid options "headerTemplate" for print to pdf. Must be string'
+                );
+            }
+            $pdfOptions['headerTemplate'] = $options['headerTemplate'];
+        }
+    
+        // option footerTemplate
+        if (array_key_exists('footerTemplate', $options)) {
+            // footerTemplate requires type to be string
+            if (!is_string($options['footerTemplate'])) {
+                throw new \InvalidArgumentException(
+                    'Invalid options "footerTemplate" for print to pdf. Must be string'
+                );
+            }
+            $pdfOptions['footerTemplate'] = $options['footerTemplate'];
         }
 
         // option paperWidth
@@ -585,7 +618,7 @@ class Page
 
         // option marginLeft
         if (array_key_exists('marginLeft', $options)) {
-            // marginBottom requires type to be float
+            // marginLeft requires type to be float
             if (gettype($options['marginLeft']) !== 'double') {
                 throw new \InvalidArgumentException(
                     'Invalid options "marginLeft" for print to pdf. Must be float like 1.0 or 5.4'
@@ -596,7 +629,7 @@ class Page
 
         // option marginRight
         if (array_key_exists('marginRight', $options)) {
-            // marginBottom requires type to be float
+            // marginRight requires type to be float
             if (gettype($options['marginRight']) !== 'double') {
                 throw new \InvalidArgumentException(
                     'Invalid options "marginRight" for print to pdf. Must be float like 1.0 or 5.4'
@@ -607,7 +640,7 @@ class Page
 
         // option preferCSSPageSize
         if (array_key_exists('preferCSSPageSize', $options)) {
-            // printBackground requires type to be boolean
+            // preferCSSPageSize requires type to be boolean
             if (!\is_bool($options['preferCSSPageSize'])) {
                 throw new \InvalidArgumentException(
                     'Invalid options "preferCSSPageSize" for print to pdf. Must be true or false'